requirements.txt
pyodbc
Pillow
requests
